Move non-app-specific Silicon Labs files into src/app.
