+++ examples/wifi-echo/server/esp32/main/gen/callback.h
@@ -46,7 +46,7 @@
  * @{
  */
 
-#include "../af-types.h"
+#include "af-types.h"
 //#include "hal/hal.h"
 //#include EMBER_AF_API_NETWORK_STEERING
