+++ examples/wifi-echo/server/esp32/main/gen/znet-bookkeeping.h
@@ -41,7 +41,7 @@
 
 //#include PLATFORM_HEADER
 //#include CONFIGURATION_HEADER
-#include "../af.h"
+#include "af.h"
 
 #ifdef __cplusplus
 extern "C" {
