+++ src/app/Makefile.am
@@ -23,6 +23,19 @@
 
 include $(abs_top_nlbuild_autotools_dir)/automake/pre.am
 
+# Package (e.g. 'make dist') these subdirectories
+
+DIST_SUBDIRS = \
+    chip-zcl \
+    chip-zcl-zpro \
+    clusters \
+    gen \
+    plugin \
+    util \
+    $(NULL)
+
+# Build (e.g for 'make all') these subdirectories
+
 SUBDIRS                               = \
     gen                                 \
     chip-zcl                            \
