+++ src/app/util/chip-response.cpp
