+++ src/app/util/chip-response.h
