+++ src/app/util/process-global-message.c
