+++ src/app/util/util.c
